+++ CLAUDE.md.spec.ts
@@ -86,7 +86,8 @@ THIS REPO'S OWN HARNESS (how Claude Code behaves when a CONTRIBUTOR works in-rep
 AUXILIARY PACKAGES (separate npm packages with their OWN \`package.json\`, NOT part of \`src/\`):
 
 - \`rule-enforcer/\` — \`@vigiles/rule-enforcer\`, the opt-in rule-SYNTHESIS engine + its two-stage blind-gold TRUST GATE (\`gate.js\`). Folded in from a former standalone repo; CI runs its gate (dogfood). The \`pr-to-lint-rule\` skill's engine.
-- \`report/\` — \`@vigiles/report\`, the Vite + React + shadcn audit-report UI, built to one self-contained HTML template the CLI fills. Kept out of the CLI's runtime deps.
+- \`report/\` — \`@vigiles/report\`, the Vite + Tailwind audit-report app that CONSUMES \`@vigiles/report-view\` and builds it to one self-contained HTML template the CLI fills. Kept out of the CLI's runtime deps.
+- \`packages/report-view/\` — \`@vigiles/report-view\`, the SHARED source-only report view (presentational components + AuditReport schema + band tokens + theme.css) rendered by \`report/\` (and later \`site/\` + the hosted demo) from the same AuditReport JSON. Wired as npm WORKSPACES (\`packages/*\`, \`report\`, \`site\`) — deps hoist to root, the published package's CI stays green (see \`research/report-view-and-hosted-demo.md\`).
 
 DOCS (two tiers — see the doc-tiers rule):
 
@@ -293,11 +294,13 @@ BUILD + TOOLING + GENERATED:
     "src/core/rule-catalog.ts":
       "The DYNAMIC available-rule catalog for the audit rule map — enumerateEslintCatalog/enumeratePylintCatalog EXECUTE the repo's real linter (own-repo + consent) to list every installed rule (plugins included) + enabled-state (Pylint via `--list-msgs` + SECTION-AWARE `--list-msgs-enabled`; matchable by symbol OR numeric code). parseEslintCatalog/parsePylintCatalog are the pure JSON/text→typed parses. mergeCatalogs unions a polyglot repo's catalogs KEEPING every entry (an id shared across linters is two real rules; collision routing is buildCatalogLookup's job, not the merge's). AvailableRule carries per-rule linter + code provenance; the 'documented but OFF' nudge reads off enabled-state. Tested in src/core/rule-catalog.test.ts + the real-binary src/rule-catalog-oss.test.ts",
     "src/audit-report.ts":
-      "The `AuditReport` — the VERSIONED JSON contract that IS the audit's product boundary (the monetization keystone). Everything renders FROM it: the local single-file HTML report, `audit --json` for CI, and (later) an UPLOAD to a hosted dashboard. meta{schemaVersion,tool,vigilesVersion,harness,dir,generatedAt?} + score (AuditScore) + recommendations + inventory. buildAuditReport(report, {harness, vigilesVersion, battery?}) assembles it — pure, no clock (the CLI stamps generatedAt at write time so the HTML-embedded form stays deterministic). VERSIONED + additive-only within a schemaVersion (it's the wire format between the CLI and any downstream). Mirrored by report/src/schema.ts (the report app builds independently)",
+      "The `AuditReport` — the VERSIONED JSON contract that IS the audit's product boundary (the monetization keystone). Everything renders FROM it: the local single-file HTML report, `audit --json` for CI, and (later) an UPLOAD to a hosted dashboard. meta{schemaVersion,tool,vigilesVersion,harness,dir,generatedAt?} + score (AuditScore) + recommendations + inventory. buildAuditReport(report, {harness, vigilesVersion, battery?}) assembles it — pure, no clock (the CLI stamps generatedAt at write time so the HTML-embedded form stays deterministic). VERSIONED + additive-only within a schemaVersion (it's the wire format between the CLI and any downstream). Mirrored by packages/report-view/src/schema.ts (the shared report view builds independently)",
     "src/audit-html.ts":
       "The shareable HTML audit report — ONE renderer (pure shadcn/Tailwind, no custom-CSS fallback). renderAuditHtml(report) injects the AuditReport JSON into the prebuilt React/shadcn template (dist/audit-report.template.html, built from report/ — see below); the React app runs in the reader's BROWSER, so the CLI stays runtime-dep-light and the output is still ONE offline file. injectReportData(template, report) is the pure testable core; <,>,& escaped on injection (no <script> breakout). The build GUARANTEES the template (build-report.mjs fails loud), so there's no inline fallback; if it's somehow missing the CLI skips the HTML (the JSON + terminal report don't depend on it). templatePath() resolves via __dirname (CommonJS output → no import.meta)",
+    "packages/report-view/":
+      "@vigiles/report-view — the SHARED audit report view (source-only, private, never published): the presentational React components (Report, Ring, RuleInventory, Adopt, Adoptability, Observations, ui/badge, ui/card), the AuditReport schema (src/schema.ts mirrors src/audit-report.ts — wire shape pinned by src/audit-report.test.ts), the band tokens (src/lib/band.ts), and a theme.css consumers import. Rendered by report/ (and later site/ + the hosted demo) from the SAME AuditReport JSON so browser and CLI show the SAME artifact — never a screenshot, never a duplicated/relative-cross-imported component (the landing-site skill's shared-UI invariant). Consumed as an npm WORKSPACE (root workspaces:[packages/*,report,site]; report depends on '*') so deps hoist to root node_modules and resolve with zero config; the published vigiles package's CI gates (api-surface, coverage) are unaffected (workspaces are private, excluded from its files array). A consumer must @source this package's src (Tailwind v4 ignores node_modules). See research/report-view-and-hosted-demo.md for the four CI-critical changes",
     "report/":
-      "The audit report UI — a SEPARATE package (@vigiles/report): Vite + React + Tailwind v4 + shadcn-style components, built via vite-plugin-singlefile to ONE self-contained index.html the CLI fills with an AuditReport (window.__VIGILES_DATA__ placeholder → injected JSON). Kept out of the published CLI's runtime deps (the frontend toolchain lives here; only the built template ships). PURE shadcn/Tailwind — band colors (good/warn/bad/na) registered as theme colors + used as utilities (text-good/stroke-bad/border-l-warn via literal class maps in src/lib/band.ts), no inline styles / no [var()] arbitrary; auto light/dark via prefers-color-scheme. Components are presentational (data-in-props) so a future HOSTED dashboard reuses them over the same AuditReport contract. scripts/build-report.mjs builds it + copies the template (wired into npm run build; FAILS LOUD if the build can't run — the report is part of the product, no silent degrade). report/src/schema.ts mirrors src/audit-report.ts (the package builds independently)",
+      "The audit report UI (@vigiles/report): a Vite + Tailwind v4 app whose src is now just main.tsx + index.css — it CONSUMES @vigiles/report-view (the shared components + schema + theme) and builds it via vite-plugin-singlefile to ONE self-contained index.html the CLI fills with an AuditReport (window.__VIGILES_DATA__ placeholder → injected JSON). Kept out of the published CLI's runtime deps (the frontend toolchain lives here; only the built template ships). scripts/build-report.mjs builds it + copies the template (wired into npm run build; FAILS LOUD if the build can't run — the report is part of the product, no silent degrade; its install-guard checks root node_modules since workspace deps are hoisted). index.css @sources ../../packages/report-view/src + @imports @vigiles/report-view/theme.css",
     "src/scaffold-test.ts":
       "The deterministic test-gen ENGINE (B1: free-form in, a RUNNABLE starter test out) — SKILL-INTERNAL, no CLI verb (the test-harness skill drives it; the standalone verb was demoted to trim the launch surface). scaffoldTest(input) is pure (ScaffoldInput → {path, content, kind, tier}): the deterministic counterpart to the test-harness SKILL (which picks the tier with a model), it picks the cheapest meaningful tier per kind — hook→unit (runHook), skill→eval (measureTriggerRate recall+precision), subagent→harness — emits at the surface's suggested test path (mirrors test-coverage's suggestedTestPath so a generated file stops the surface being reported untested), wires the real PUBLIC API (vigiles/unit, vigiles/testing, vigiles/spec) + the surface's own metadata, and leaves TODOs only where a human/model must supply judgement (prompts/event/values). THE TYPED-SPEC PAYOFF (the thing markdown can't generate): for a subagent the generator CONSUMES the typed contract — a parsed result() contract (ResultContract from the compiled .md's vigiles:ok/err blocks) → a deterministic assertAgentOk OUTCOME test reconstructing the real fields (NO LLM judge), and the side-effecting tools (effectSurface(tools, dialect).sideEffecting) → a generated SAFETY check (notTool for Bash, didNotWrite for Write/Edit) that asserts the agent's 'hole' stays in its lane (the auto-derive-interceptTools nugget from effect-boundary-design.md, realized at scaffold time). So a spec WRITES its own outcome+safety tests. formatScaffolds renders the summary",
     "src/scaffold-test.test.ts":
