+++ packages/report-view/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@vigiles/report-view",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "description": "Shared audit report view — the presentational React components + the AuditReport schema + band tokens, rendered by report/, site/, and the hosted demo from the same AuditReport JSON. Source-only (bundler-consumed via workspace symlink); no build step, never published.",
+  "exports": {
+    ".": "./src/index.ts",
+    "./theme.css": "./src/theme.css"
+  },
+  "peerDependencies": {
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
+  },
+  "dependencies": {
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "lucide-react": "^0.469.0",
+    "tailwind-merge": "^2.6.0"
+  }
+}
