+++ packages/report-view/src/lib/band.ts
