+++ research/CLAUDE.md.spec.ts
@@ -47,6 +47,7 @@ export default claude({
     "research/audit-adoption-ux.md": `[shipped] Design record for how audit report drives spec creation (init flow); default creates specs, no browser file writes`,
     "research/audit-eval-thickening-and-gate.md": `[shipped] Records audit's behavioral tier expanding to three evals (fire/collide/hold) plus the adversarial-gate eval build`,
     "research/audit-lighthouse-design.md": `[shipped] Locked design for vigiles audit as "Lighthouse for your harness" — rings, HTML report, battery, --deep tier`,
+    "research/report-view-and-hosted-demo.md": `[active] The shared @vigiles/report-view package (SHIPPED via npm workspaces — the four CI-critical changes: root lockfile regen, pages.yml site build, build-report install-guard, sub-lockfile deletion; + the Tailwind @source gotcha) + the Stage-2 hosted "Grade a repo" demo design of record (Fable brief: zero-click baked report, honest gate/tease, edge cases, funnel)`,
     "research/audit-harness-dx.md": `[deferred] Multi-harness audit DX design (detection: AGENTS.md-≠-Codex per the AAIF standard + mirror-collapse; shared-vs-dialect-specific factoring; perf local/browser/LLM; metered-API cost warning) — DEFERRED 2026-07-21: audit is CLAUDE-CODE-FOCUSED for now. The preserved record + research for when we revisit`,
     "research/audit-serve-design.md": `[idea] Design for an opt-in audit --serve local server enabling one-click spec creation from the HTML report`,
     "research/audit-wow-ideas.md": `[active] Research on what audit findings create genuine "wow" tension beyond saturated markdown-linter checks`,
